feat(cmd): 实现统一命令调度框架 新增统一命令调度框架，整合 IR/ESP-NOW/CAN/UART/WebUI 多链路命令收发： 核心框架: - app_cmd.c/h: 命令注册、路由、响应机制 - app_cmd_handlers.c/h: 12 个内置命令处理器 - app_cmd_test.c/h: 自动测试框架 命令类型: - 系统命令: QUERY_STATUS/VERSION/STATS、REBOOT、CONFIG - 设备控制: LED_CTRL/EFFECT/BRIGHTNESS (WS2812) - 显示控制: OLED_CLEAR/TEXT/REFRESH - 输入设备: KEY_QUERY/EVENT - 桥接命令: IR↔ESP-NOW、CAN↔ESP-NOW、IR↔CAN - 用户自定义: 0xF0-0xFF 特性: - 多链路支持: IR、ESP-NOW、CAN、UART、WebUI、本地 - 动态注册: 运行时注册/注销命令处理器 - 统计功能: 记录各链路收发次数、成功/失败 - 线程安全: FreeRTOS 互斥锁保护 文档: - doc/cmd_design.md: 完整设计文档

+++ main/app/app_devtest.c
@@ -218,13 +218,13 @@ static void twai_test_tx_task(void *arg) {
     if (ret == ESP_OK) {
       s_twai_tx_count++;
       if (s_twai_tx_count % 100 == 0) {
-        ESP_LOGI(TAG, "TWAI TX #%lu OK", (unsigned long)s_twai_tx_count);
+        ESP_LOGI(TAG, "TWAI TX #%" PRIu32 " OK", s_twai_tx_count);
       }
     } else {
       ESP_LOGW(TAG, "TWAI TX fail: %d (0x%x)", ret, ret);
       twai_status_info_t st;
       if (bsp_twai_get_status(&st) == ESP_OK) {
-        ESP_LOGW(TAG, "TWAI state=%d tx_err=%lu rx_err=%lu",
+        ESP_LOGW(TAG, "TWAI state=%d tx_err=%" PRIu32 " rx_err=%" PRIu32,
                  st.state, st.tx_error_counter, st.rx_error_counter);
       }
     }
